requirements.txt
aiohttp==3.8.6
aiosignal==1.3.1
anyio==4.0.0
async-timeout==4.0.3
attrs==23.1.0
blinker==1.7.0
build==1.0.3
cachecontrol==0.13.1
cattrs==22.2.0
certifi==2023.7.22
cffi==1.16.0
charset-normalizer==3.3.2
cleo==2.1.0
click==8.1.7
crashtest==0.4.1
cryptography==41.0.5
distlib==0.3.7
dulwich==0.21.6
exceptiongroup==1.1.3
fastjsonschema==2.18.1
filelock==3.13.1
flask==3.0.0
frozenlist==1.4.0
gunicorn==21.2.0
h11==0.14.0
highrise-bot-sdk==23.3.4
httpcore==1.0.1
httpx==0.25.1
idna==3.4
importlib-metadata==6.8.0
installer==0.7.0
itsdangerous==2.1.2
jaraco.classes==3.3.0
jeepney==0.8.0
Jinja2==3.1.2
keyring==24.2.0
MarkupSafe==2.1.3
more-itertools==10.1.0
msgpack==1.0.7
multidict==6.0.4
packaging==23.2
pendulum==2.1.2
pexpect==4.8.0
pkginfo==1.9.6
platformdirs==3.11.0
poetry==1.7.0
poetry-core==1.8.1
poetry-plugin-export==1.6.0
ptyprocess==0.7.0
pycparser==2.21
pyproject-hooks==1.0.0
python-dateutil==2.8.2
pytzdata==2020.1
quattro==22.2.0
rapidfuzz==3.5.2
requests==2.31.0
requests-toolbelt==1.0.0
SecretStorage==3.3.3
shellingham==1.5.4
six==1.16.0
sniffio==1.3.0
tomli==2.0.1
tomlkit==0.12.2
trove-classifiers==2023.11.7
typing-extensions==3.10.0.2
urllib3==2.0.7
virtualenv==20.24.6
werkzeug==3.0.1
yarl==1.9.2
zipp==3.17.0
